commands: merge commands.rs and control.rs in a commands/ mod
This doesn't entirely fixes the circular dependency between mod.rs and
utils.rs, but at least contains it in a single module.
